feat: parking statistics by timestamps
todo: stats only by date and only by time

+++ src/main/java/pl/wrapper/parking/facade/dto/ParkingStatsResponse.java
@@ -2,5 +2,5 @@ package pl.wrapper.parking.facade.dto;
 
 import java.time.LocalDateTime;
 
-public record ParkingStatsResponse(int totalUsage, double averageAvailability, LocalDateTime peakOccupancyAt) {
+public record ParkingStatsResponse(long totalUsage, double averageAvailability, LocalDateTime peakOccupancyAt) {
 }
